random-image-generator/requirements.txt
pillow==4.3.0
requests==2.18.4
